Desktop/Projet/P3000/Application/requirements.txt
Django==5.1.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-storages==1.14.2
boto3==1.34.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
dj-database-url==2.1.0
whitenoise==6.6.0
PyMuPDF==1.23.8
Pillow==10.2.0
requests==2.31.0
python-decouple==3.8
PyJWT==2.8.0
